Bugfixes for Dockerfile
I need to write credentials into the right place in a derivative
dockerfile, rather than just mounting a volume directly, so now
the user which this runs as can write to its own home directory.

Also changed the `--config` command-line argume to accept multiple
config files, which will be merged in the order listed.